+++ src/shapes/ActiveSelection.spec.ts
@@ -1,9 +1,18 @@
+import { FitContentLayout } from '../LayoutManager';
 import { Canvas } from '../canvas/Canvas';
 import { ActiveSelection } from './ActiveSelection';
 import { Group } from './Group';
 import { FabricObject } from './Object/FabricObject';
 
 describe('ActiveSelection', () => {
+  it('should set the layoutManager in the constructor', () => {
+    const activeSelection = new ActiveSelection();
+    expect(activeSelection.layoutManager).toBeDefined();
+    expect(activeSelection.layoutManager?.strategy).toBeInstanceOf(
+      FitContentLayout
+    );
+  });
+
   it('clearing active selection objects resets transform', () => {
     const obj = new FabricObject({
       left: 100,
@@ -88,7 +97,7 @@ describe('ActiveSelection', () => {
   });
 
   it('`setActiveObject` should update the active selection ref on canvas if it changed', () => {
-    const canvas = new Canvas(null);
+    const canvas = new Canvas();
     const obj1 = new FabricObject();
     const obj2 = new FabricObject();
     canvas.add(obj1, obj2);
@@ -105,18 +114,76 @@ describe('ActiveSelection', () => {
     expect(spy).not.toHaveBeenCalled();
   });
 
-  it('transferring an object between active selections keeps its owning group', () => {
+  test('adding and removing an object belonging to a group', () => {
+    const object = new FabricObject();
+    const group = new Group([object]);
+    const canvas = new Canvas();
+    const activeSelection = canvas.getActiveSelection();
+
+    const eventsSpy = jest.spyOn(object, 'fire');
+    const removeSpy = jest.spyOn(group, 'remove');
+    const exitSpy = jest.spyOn(group, '_exitGroup');
+    const enterSpy = jest.spyOn(activeSelection, 'enterGroup');
+
+    expect(object.group).toBe(group);
+    expect(object.parent).toBe(group);
+    expect(object.canvas).toBeUndefined();
+
+    activeSelection.add(object);
+    expect(object.group).toBe(activeSelection);
+    expect(object.parent).toBe(group);
+    expect(object.canvas).toBe(canvas);
+    expect(removeSpy).not.toBeCalled();
+    expect(exitSpy).toBeCalledWith(object);
+    expect(enterSpy).toBeCalledWith(object, true);
+    expect(eventsSpy).toHaveBeenNthCalledWith(1, 'added', {
+      target: activeSelection,
+    });
+
+    activeSelection.remove(object);
+    expect(eventsSpy).toHaveBeenNthCalledWith(2, 'removed', {
+      target: activeSelection,
+    });
+    expect(object.group).toBe(group);
+    expect(object.parent).toBe(group);
+    expect(object.canvas).toBeUndefined();
+  });
+
+  test('transferring an object between active selections', () => {
     const object = new FabricObject();
     const group = new Group([object]);
     const activeSelection1 = new ActiveSelection([object]);
     const activeSelection2 = new ActiveSelection();
     expect(object.group).toBe(activeSelection1);
-    expect(object.getParent(true)).toBe(group);
+    expect(object.parent).toBe(group);
+
+    const eventsSpy = jest.spyOn(object, 'fire');
+    const removeSpy = jest.spyOn(activeSelection1, 'remove');
+
+    Object.entries({
+      object,
+      group,
+      activeSelection1,
+      activeSelection2,
+    }).forEach(([key, obj]) => jest.spyOn(obj, 'toJSON').mockReturnValue(key));
+
     activeSelection2.add(object);
     expect(object.group).toBe(activeSelection2);
-    expect(object.getParent(true)).toBe(group);
+    expect(object.parent).toBe(group);
+    expect(removeSpy).toBeCalledWith(object);
+    expect(eventsSpy).toHaveBeenNthCalledWith(1, 'removed', {
+      target: activeSelection1,
+    });
+    expect(eventsSpy).toHaveBeenNthCalledWith(2, 'added', {
+      target: activeSelection2,
+    });
+
     activeSelection2.removeAll();
     expect(object.group).toBe(group);
-    expect(object.getParent(true)).toBe(group);
+    expect(object.parent).toBe(group);
+    expect(eventsSpy).toHaveBeenNthCalledWith(3, 'removed', {
+      target: activeSelection2,
+    });
+    expect(eventsSpy).toBeCalledTimes(3);
   });
 });